Reorg: Move utils to subpackage
