+++ src/main/java/com/ginsberg/gatherers4j/CrossGatherer.java
@@ -26,7 +26,7 @@ import java.util.stream.Gatherer;
 import java.util.stream.Stream;
 import java.util.stream.StreamSupport;
 
-import static com.ginsberg.gatherers4j.GathererUtils.mustNotBeNull;
+import static com.ginsberg.gatherers4j.util.GathererUtils.mustNotBeNull;
 
 public class CrossGatherer {
