+++ src/main/java/com/ginsberg/gatherers4j/DropLastGatherer.java
@@ -16,6 +16,7 @@
 
 package com.ginsberg.gatherers4j;
 
+import com.ginsberg.gatherers4j.util.CircularBuffer;
 import org.jspecify.annotations.Nullable;
 
 import java.util.function.Supplier;
