+++ src/main/java/com/ginsberg/gatherers4j/FrequencyGatherer.java
@@ -27,7 +27,7 @@ import java.util.function.BinaryOperator;
 import java.util.function.Supplier;
 import java.util.stream.Gatherer;
 
-import static com.ginsberg.gatherers4j.GathererUtils.mustNotBeNull;
+import static com.ginsberg.gatherers4j.util.GathererUtils.mustNotBeNull;
 
 public class FrequencyGatherer<INPUT extends @Nullable Object>
         implements Gatherer<INPUT, FrequencyGatherer.State<INPUT>, WithCount<INPUT>> {
